refactor(app): 재게시 액션 책임을 모듈로 분리한다
