+++ apps/app/src/components/post/useRepostFailureToast.ts
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
 import { useToast } from '@/components/ui/ToastProvider';
-import type { RepostActionFailure, RepostActionKind } from './PostActionBar';
+import type { RepostActionFailure, RepostActionKind } from './RepostAction';
 
 const messages: Record<RepostActionKind, string> = {
   create: '재게시하지 못했습니다. 잠시 후 다시 시도해 주세요.',
